Get Currency from string (+ refactoring)
